pyproject.toml
[project]
name = "agent-workflow-server"
version = "0.2.1"
description = ""
authors = [
    {name = "Janos Sarusi-Kis", email = "janossk@cisco.com"},
    {name = "Marco Trinelli ", email = "mtrinell@cisco.com"}
]

readme = "README.md"
requires-python = ">=3.12,<4"

dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "langgraph (>=0.2.60,<0.4.0)",
    "llama-index (>=0.12.0,<0.13.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "openapi-spec-validator (>=0.7.1,<0.8.0)",
    "pydantic-ai (>=0.0.53)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "agent_workflow_server.main:start"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-mock = "^3.14.0"
ruff = "^0.11.0"
pytest-asyncio = "^0.25.3"

[tool.ruff]
lint.select = [ "E", "F", "I", "TID251"]
lint.ignore = [ "E501" ]
indent-width = 4
exclude = ["src/agent_workflow_server/generated/*"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests"
]
asyncio_default_fixture_loop_scope="function"